:sparkles:feat:등급 배치 끝

+++ src/main/java/colorful/starbucks/batch/MemberLevelPolicy.java
@@ -3,7 +3,7 @@ package colorful.starbucks.batch;
 import colorful.starbucks.member.domain.MemberLevel;
 
 public class MemberLevelPolicy {
-    public static MemberLevel calculateLevel(Integer totalAmount) {
+    public static MemberLevel calculateLevel(Long totalAmount) {
         if (totalAmount >= 1000000) {
             return MemberLevel.BLACK;
         } else if (totalAmount >= 500000) {
