+++ src/main/java/colorful/starbucks/batch/dto/MemberLevelUpdateDto.java
@@ -9,9 +9,9 @@ import lombok.NoArgsConstructor;
 public class MemberLevelUpdateDto {
     private String memberUuid;
     private MemberLevel memberLevel;
-    private Integer totalAmount;
+    private Long totalAmount;
 
-    public MemberLevelUpdateDto(String memberUuid, MemberLevel memberLevel, Integer totalAmount) {
+    public MemberLevelUpdateDto(String memberUuid, MemberLevel memberLevel, Long totalAmount) {
         this.memberUuid = memberUuid;
         this.memberLevel = memberLevel;
         this.totalAmount = totalAmount;
